# docs/swagger.yaml
basePath: /
definitions:
  api.AddEmployeeResponse:
    properties:
      Id:
        type: integer
    type: object
  api.AddPatchEmployee:
    properties:
      CompanyId:
        type: integer
      Department:
        properties:
          Name:
            type: string
          Phone:
            type: string
        type: object
      Name:
        type: string
      Passport:
        properties:
          Number:
            type: string
          Type:
            type: string
        type: object
      Phone:
        type: string
      Surname:
        type: string
    type: object
  api.Employee:
    properties:
      CompanyId:
        type: integer
      Department:
        properties:
          Name:
            type: string
          Phone:
            type: string
        type: object
      Id:
        type: integer
      Name:
        type: string
      Passport:
        properties:
          Number:
            type: string
          Type:
            type: string
        type: object
      Phone:
        type: string
      Surname:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      Code:
        type: integer
      Description:
        type: string
    type: object
  api.HealthResponse:
    properties:
      Status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления сотрудниками компаний.
  title: Employee Service API
  version: "1.0"
paths:
  /company/employees:
    get:
      description: Возвращает сотрудников по идентификатору компании.
      parameters:
      - description: Идентификатор компании
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Employee'
            type: array
        "204":
          description: Сотрудники не найдены
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Получить сотрудников компании
      tags:
      - employees
  /department/employees:
    get:
      description: Возвращает сотрудников по названию отдела.
      parameters:
      - description: Название отдела
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Employee'
            type: array
        "204":
          description: Сотрудники не найдены
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Получить сотрудников отдела
      tags:
      - employees
  /employees:
    delete:
      description: Удаляет сотрудника по идентификатору.
      parameters:
      - description: Идентификатор сотрудника
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сотрудник удалён
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Удалить сотрудника
      tags:
      - employees
    patch:
      consumes:
      - application/json
      description: Частично обновляет сотрудника по идентификатору.
      parameters:
      - description: Идентификатор сотрудника
        in: query
        name: id
        required: true
        type: integer
      - description: Изменяемые поля сотрудника
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/api.AddPatchEmployee'
      produces:
      - application/json
      responses:
        "200":
          description: Сотрудник обновлён
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Обновить сотрудника
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Создаёт сотрудника и возвращает его идентификатор.
      parameters:
      - description: Данные сотрудника
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/api.AddPatchEmployee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AddEmployeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Добавить сотрудника
      tags:
      - employees
  /health:
    get:
      description: Проверяет доступность API и соединение с PostgreSQL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Проверить состояние сервиса
      tags:
      - health
schemes:
- http
swagger: "2.0"